feat(tests): added ARIMA and Exp. Smoothing to probabilistic tests
